requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
scikit-learn>=1.3
langchain>=0.1.0
langchain-community>=0.0.1
langchain-huggingface>=0.0.1
transformers>=4.35
torch>=2.0
gunicorn>=21.0
python-dotenv>=1.0
